+++ src/IV1350_seminar_3/model/ItemTest.java
@@ -1,4 +1,4 @@
-package IV1350_seminar_3.tests;
+package IV1350_seminar_3.model;
 
 import IV1350_seminar_3.model.Item;
 import org.junit.jupiter.api.AfterEach;
