+++ src/IV1350_seminar_3/model/NoItemIDExceptionTest.java
@@ -1,4 +1,4 @@
-package IV1350_seminar_3.tests;
+package IV1350_seminar_3.model;
 
         import IV1350_seminar_3.integration.ItemDatabase;
         import IV1350_seminar_3.model.Item;
