+++ src/IV1350_seminar_3/view/View.java
@@ -20,8 +20,11 @@ public class View {
     private int thirdItemID = 404;
     private int thirdItemQuantity = 3;
 
-    private int fourthItemID = 3;
-    private int fourthItemQuantity = -4;
+    private int fourthItemID = 1337;
+    private int fourthItemQuantity = 1;
+
+    private int fifthItemID = 1;
+    private int fifthItemQuantity = -1;
 
 
     /**
@@ -62,6 +65,8 @@ public class View {
         ItemDTO fourthItemDTO = controller.itemScan(fourthItemID, fourthItemQuantity);
         itemToPrint(fourthItemDTO);
 
+        ItemDTO fifthItemDTO = controller.itemScan(fifthItemID, fifthItemQuantity);
+        itemToPrint(fifthItemDTO);
 
         float totalPrice = controller.endSale();
         controller.depositAmountPaid(totalPrice);
